requirements.txt
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask==2.1.2
greenlet==1.1.3.post0
importlib-metadata==5.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
pip==22.3
PyMySQL==1.0.2
pyrsistent==0.18.1
pytz==2022.5
PyYAML==6.0
setuptools==65.5.0
six==1.16.0
SQLAlchemy==1.4.42
Werkzeug==2.1.2
wheel==0.37.1
zipp==3.9.0

requests~=2.28.1
yfinance==0.1.80
pandas~=1.5.1